Refactor layout into LayoutWrapper component and clean up root.tsx
